services/auth_service/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
pydantic
python-jose
tenacity
